pyproject.toml
[project]
name = "autosearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "googlesearch-python",
]
